[feat] add flags to countries
